pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "project-qa-assistant"
version = "0.1.0"
description = "Code-aware RAG assistant for go-commerce"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.8.0",
  "python-dotenv>=1.0.1",
  "typer>=0.12.3",
  "rich>=13.7.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.5.0",
  "mypy>=1.10.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
